Refactor sender module, fixes #140
