+++ tekore/_sender/extending.py
@@ -1,199 +1,13 @@
 import time
 import asyncio
 
-from abc import ABC, abstractmethod
+from abc import ABC
 from typing import Optional
-from warnings import warn
 from collections import deque
 from urllib.parse import urlencode
 
-from httpx import AsyncClient
-from requests import Request, Response, Session
-
-
-class Sender(ABC):
-    """
-    Sender interface for requests.
-    """
-    @abstractmethod
-    def send(self, request: Request) -> Response:
-        """
-        Prepare and send a request.
-
-        Parameters
-        ----------
-        request
-            :class:`Request` to send
-        """
-
-    @property
-    @abstractmethod
-    def is_async(self) -> bool:
-        """
-        :class:`True` if the sender is asynchronous, :class:`False` otherwise.
-        """
-
-
-class SyncSender(Sender, ABC):
-    """
-    Synchronous request sender base class.
-    """
-    @property
-    def is_async(self) -> bool:
-        """
-        Sender asynchronicity, always :class:`False`.
-        """
-        return False
-
-
-class AsyncSender(Sender, ABC):
-    """
-    Asynchronous request sender base class.
-    """
-    @property
-    def is_async(self) -> bool:
-        """
-        Sender asynchronicity, always :class:`True`.
-        """
-        return True
-
-
-class TransientSender(SyncSender):
-    """
-    Create a new session for each request.
-
-    Parameters
-    ----------
-    requests_kwargs
-        keyword arguments for :meth:`requests.Session.send`
-    """
-    def __init__(self, **requests_kwargs):
-        from tekore import default_requests_kwargs
-        self.requests_kwargs = requests_kwargs or default_requests_kwargs
-
-    def send(self, request: Request) -> Response:
-        with Session() as sess:
-            prepared = sess.prepare_request(request)
-            return sess.send(prepared, **self.requests_kwargs)
-
-
-class AsyncTransientSender(AsyncSender):
-    """
-    Create a new asynchronous client for each request.
-
-    Parameters
-    ----------
-    httpx_kwargs
-        keyword arguments for :meth:`httpx.AsyncClient.request`
-    """
-    def __init__(self, **httpx_kwargs):
-        from tekore import default_httpx_kwargs
-        self.httpx_kwargs = httpx_kwargs or default_httpx_kwargs
-
-    async def send(self, request: Request) -> Response:
-        async with AsyncClient() as client:
-            return await client.request(
-                request.method,
-                request.url,
-                data=request.data or None,
-                params=request.params or None,
-                headers=request.headers,
-                **self.httpx_kwargs,
-            )
-
-
-class SingletonSender(SyncSender):
-    """
-    Use one session for all instances and requests.
-
-    Parameters
-    ----------
-    requests_kwargs
-        keyword arguments for :meth:`requests.Session.send`
-    """
-    session = Session()
-
-    def __init__(self, **requests_kwargs):
-        from tekore import default_requests_kwargs
-        self.requests_kwargs = requests_kwargs or default_requests_kwargs
-
-    def send(self, request: Request) -> Response:
-        prepared = SingletonSender.session.prepare_request(request)
-        return SingletonSender.session.send(prepared, **self.requests_kwargs)
-
-
-class AsyncSingletonSender(AsyncSender):
-    """
-    Use one client for all instances and requests.
-
-    Parameters
-    ----------
-    httpx_kwargs
-        keyword arguments for :meth:`httpx.AsyncClient.request`
-    """
-    client = AsyncClient()
-
-    def __init__(self, **httpx_kwargs):
-        from tekore import default_httpx_kwargs
-        self.httpx_kwargs = httpx_kwargs or default_httpx_kwargs
-
-    async def send(self, request: Request) -> Response:
-        return await AsyncSingletonSender.client.request(
-            request.method,
-            request.url,
-            data=request.data or None,
-            params=request.params or None,
-            headers=request.headers,
-            **self.httpx_kwargs,
-        )
-
-
-class PersistentSender(SyncSender):
-    """
-    Use a per-instance session to send requests.
-
-    Parameters
-    ----------
-    session
-        :class:`requests.Session` to use when sending requests
-    requests_kwargs
-        keyword arguments for :meth:`requests.Session.send`
-    """
-    def __init__(self, session: Session = None, **requests_kwargs):
-        from tekore import default_requests_kwargs
-        self.requests_kwargs = requests_kwargs or default_requests_kwargs
-        self.session = session or Session()
-
-    def send(self, request: Request) -> Response:
-        prepared = self.session.prepare_request(request)
-        return self.session.send(prepared, **self.requests_kwargs)
-
-
-class AsyncPersistentSender(AsyncSender):
-    """
-    Use a per-instance client to send requests asynchronously.
-
-    Parameters
-    ----------
-    session
-        :class:`httpx.AsyncClient` to use when sending requests
-    httpx_kwargs
-        keyword arguments for :meth:`httpx.AsyncClient.request`
-    """
-    def __init__(self, client: AsyncClient = None, **httpx_kwargs):
-        from tekore import default_httpx_kwargs
-        self.httpx_kwargs = httpx_kwargs or default_httpx_kwargs
-        self.client = client or AsyncClient()
-
-    async def send(self, request: Request) -> Response:
-        return await self.client.request(
-            request.method,
-            request.url,
-            data=request.data or None,
-            params=request.params or None,
-            headers=request.headers,
-            **self.httpx_kwargs,
-        )
+from requests import Request, Response
+from .concrete import Sender
 
 
 class ExtendingSender(Sender, ABC):
@@ -484,51 +298,3 @@ class CachingSender(ExtendingSender):
         fresh = await self.sender.send(request)
         async with self._lock:
             return self._handle_fresh(request, fresh, cached)
-
-
-def new_default_sender() -> Sender:
-    from tekore import default_sender_instance, default_sender_type
-    return default_sender_instance or default_sender_type()
-
-
-class SenderConflictWarning(RuntimeWarning):
-    """
-    Sender arguments to a client are in conflict.
-    """
-
-
-class Client:
-    """
-    Base class for clients.
-
-    Parameters
-    ----------
-    sender
-        request sender - If not specified, using :attr:`default_sender_instance`
-        is attempted first, then :attr:`default_sender_type` is instantiated.
-    asynchronous
-        synchronicity requirement - If specified, overrides passed
-        sender and defaults if they are in conflict and instantiates
-        a transient sender of the requested type
-    """
-    def __init__(self, sender: Optional[Sender], asynchronous: bool = None):
-        new_sender = sender or new_default_sender()
-
-        if new_sender.is_async and asynchronous is False:
-            new_sender = TransientSender()
-        elif not new_sender.is_async and asynchronous is True:
-            new_sender = AsyncTransientSender()
-
-        self.sender = new_sender
-
-        if sender is not None and new_sender.is_async != sender.is_async:
-            msg = f'\n{type(sender)} passed but asynchronous={asynchronous}!'
-            msg += '\nA sender was instantiated according to `asynchronous`.'
-            warn(msg, SenderConflictWarning, stacklevel=3)
-
-    def _send(self, request: Request) -> Response:
-        return self.sender.send(request)
-
-    @property
-    def is_async(self) -> bool:
-        return self.sender.is_async
